requirements.txt
# Core
feedparser>=6.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# LLM classification — OpenAI-compatible client (supports GLM, DeepSeek, Gemini, Groq, etc.)
openai>=1.30.0

# Data processing
pandas>=2.0.0

# Semantic search (optional)
# sentence-transformers>=2.2.0
# numpy>=1.24.0

# HTTP server for local preview
# (built-in: python3 -m http.server)